Fix compile error on certain GPU models

+++ dpvslam/fastba/ba_cuda.cu
@@ -332,7 +332,7 @@ __global__ void reprojection_residuals_and_hessian(
 
       }
 
-      atomicAdd(&r_total[0],  w * r * r);
+      atomicAdd(&r_total[0],  (double)(w * r * r));
 
       adjSE3(tij, qij, Jj, Ji);
